remove multiple deploys
